Create AppSearchButton.vue
